backend/requirements.txt
alembic==1.13.2
black==24.8.0
blinker==1.8.2
click==8.1.7
Faker==28.4.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
graphql-core==2.3.2
graphql-server-core==1.2.0
greenlet==3.1.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
promise==2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Rx==1.6.3
six==1.16.0
SQLAlchemy==2.0.34
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.1
